move Print trait and implementations into a separate module;
