+++ ergotree-ir/src/pretty_printer/print.rs
@@ -1,0 +1,227 @@
+use thiserror::Error;
+
+use crate::mir::bin_op::BinOp;
+use crate::mir::block::BlockValue;
+use crate::mir::coll_append::Append;
+use crate::mir::coll_by_index::ByIndex;
+use crate::mir::constant::Constant;
+use crate::mir::expr::Expr;
+use crate::mir::global_vars::GlobalVars;
+use crate::mir::val_def::ValDef;
+use crate::mir::val_use::ValUse;
+use crate::source_span::SourceSpan;
+use crate::source_span::Spanned;
+
+use super::Printer;
+
+/// Print error
+#[allow(missing_docs)]
+#[derive(PartialEq, Eq, Debug, Clone, Error)]
+pub enum PrintError {
+    #[error("fmt error: {0:?}")]
+    FmtError(#[from] std::fmt::Error),
+}
+
+/// Print trait for Expr that sets the source span for the resulting Expr
+pub trait Print {
+    /// Print the expression and return the resulting expression with source span
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError>;
+}
+
+impl Print for BlockValue {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        let offset = w.current_pos();
+        writeln!(w, "{{")?;
+        w.inc_ident();
+        let indent = w.get_indent();
+        let mut items = Vec::new();
+        for item in &self.items {
+            write!(w, "{:indent$}", "", indent = indent)?;
+            items.push(item.print(w)?);
+            writeln!(w)?;
+        }
+        // indent for result
+        write!(w, "{:indent$}", "", indent = indent)?;
+        let res = self.result.print(w)?;
+        w.dec_ident();
+        writeln!(w, "\n}}")?;
+        let length = w.current_pos() - offset;
+        Ok(Spanned {
+            source_span: SourceSpan { offset, length },
+            expr: BlockValue {
+                items,
+                result: Box::new(res),
+            },
+        }
+        .into())
+    }
+}
+
+impl Print for ValDef {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        let offset = w.current_pos();
+        write!(w, "val v{} = ", self.id)?;
+        let rhs = self.rhs.print(w)?;
+        let length = w.current_pos() - offset;
+        Ok(Spanned {
+            source_span: SourceSpan { offset, length },
+            expr: ValDef {
+                id: self.id,
+                rhs: Box::new(rhs),
+            },
+        }
+        .into())
+    }
+}
+
+impl Print for Constant {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        write!(w, "{:?}", self.v)?;
+        Ok(self.clone().into())
+    }
+}
+
+impl Print for ValUse {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        write!(w, "v{}", self.val_id)?;
+        Ok(self.clone().into())
+    }
+}
+
+impl Print for Append {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        let offset = w.current_pos();
+        let input = self.input.print(w)?;
+        write!(w, ".append(")?;
+        let col_2 = self.col_2.print(w)?;
+        write!(w, ")")?;
+        let length = w.current_pos() - offset;
+        Ok(Spanned {
+            source_span: SourceSpan { offset, length },
+            expr: Append {
+                input: Box::new(input),
+                col_2: Box::new(col_2),
+            },
+        }
+        .into())
+    }
+}
+
+impl Print for BinOp {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        let offset = w.current_pos();
+        let left = self.left.print(w)?;
+        write!(w, " {} ", self.kind)?;
+        let right = self.right.print(w)?;
+        let length = w.current_pos() - offset;
+        // dbg!(offset, length);
+        Ok(Spanned {
+            source_span: SourceSpan { offset, length },
+            expr: BinOp {
+                kind: self.kind,
+                left: Box::new(left),
+                right: Box::new(right),
+            },
+        }
+        .into())
+    }
+}
+
+impl Print for GlobalVars {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        write!(w, "{}", self)?;
+        Ok(self.clone().into())
+    }
+}
+
+impl Print for ByIndex {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        let input = self.input.print(w)?;
+        let offset = w.current_pos();
+        write!(w, "(")?;
+        let index = self.index.print(w)?;
+        write!(w, ")")?;
+        let length = w.current_pos() - offset;
+        #[allow(clippy::unwrap_used)] // we only added spans
+        Ok(Spanned {
+            source_span: SourceSpan { offset, length },
+            expr: ByIndex::new(input, index, self.default.clone()).unwrap(),
+        }
+        .into())
+    }
+}
+
+#[allow(clippy::todo)]
+#[allow(clippy::panic)]
+impl Print for Expr {
+    fn print(&self, w: &mut dyn Printer) -> Result<Expr, PrintError> {
+        match self {
+            Expr::Append(v) => v.expr().print(w),
+            Expr::BlockValue(v) => v.expr().print(w),
+            Expr::ValDef(v) => v.expr().print(w),
+            Expr::ValUse(v) => v.print(w),
+            Expr::Const(v) => v.print(w),
+            Expr::BinOp(v) => v.expr().print(w),
+            Expr::GlobalVars(v) => v.print(w),
+            Expr::ByIndex(v) => v.expr().print(w),
+            Expr::ConstPlaceholder(_) => Ok(self.clone()),
+            Expr::SubstConstants(_) => todo!(),
+            Expr::ByteArrayToLong(_) => todo!(),
+            Expr::ByteArrayToBigInt(_) => todo!(),
+            Expr::LongToByteArray(_) => todo!(),
+            Expr::Collection(_) => todo!(),
+            Expr::Tuple(_) => todo!(),
+            Expr::CalcBlake2b256(_) => todo!(),
+            Expr::CalcSha256(_) => todo!(),
+            Expr::Context => todo!(),
+            Expr::Global => todo!(),
+            Expr::FuncValue(_) => todo!(),
+            Expr::Apply(_) => todo!(),
+            Expr::MethodCall(_) => todo!(),
+            Expr::PropertyCall(_) => todo!(),
+            Expr::If(_) => todo!(),
+            Expr::And(_) => todo!(),
+            Expr::Or(_) => todo!(),
+            Expr::Xor(_) => todo!(),
+            Expr::Atleast(_) => todo!(),
+            Expr::LogicalNot(_) => todo!(),
+            Expr::Negation(_) => todo!(),
+            Expr::BitInversion(_) => todo!(),
+            Expr::OptionGet(_) => todo!(),
+            Expr::OptionIsDefined(_) => todo!(),
+            Expr::OptionGetOrElse(_) => todo!(),
+            Expr::ExtractAmount(_) => todo!(),
+            Expr::ExtractRegisterAs(_) => todo!(),
+            Expr::ExtractBytes(_) => todo!(),
+            Expr::ExtractBytesWithNoRef(_) => todo!(),
+            Expr::ExtractScriptBytes(_) => todo!(),
+            Expr::ExtractCreationInfo(_) => todo!(),
+            Expr::ExtractId(_) => todo!(),
+            Expr::SizeOf(_) => todo!(),
+            Expr::Slice(_) => todo!(),
+            Expr::Fold(_) => todo!(),
+            Expr::Map(_) => todo!(),
+            Expr::Filter(_) => todo!(),
+            Expr::Exists(_) => todo!(),
+            Expr::ForAll(_) => todo!(),
+            Expr::SelectField(_) => todo!(),
+            Expr::BoolToSigmaProp(_) => todo!(),
+            Expr::Upcast(_) => todo!(),
+            Expr::Downcast(_) => todo!(),
+            Expr::CreateProveDlog(_) => todo!(),
+            Expr::CreateProveDhTuple(_) => todo!(),
+            Expr::SigmaPropBytes(_) => todo!(),
+            Expr::DecodePoint(_) => todo!(),
+            Expr::SigmaAnd(_) => todo!(),
+            Expr::SigmaOr(_) => todo!(),
+            Expr::GetVar(_) => todo!(),
+            Expr::DeserializeRegister(_) => todo!(),
+            Expr::DeserializeContext(_) => todo!(),
+            Expr::MultiplyGroup(_) => todo!(),
+            Expr::Exponentiate(_) => todo!(),
+            Expr::XorOf(_) => todo!(),
+            Expr::TreeLookup(_) => todo!(),
+            Expr::CreateAvlTree(_) => todo!(),
+        }
+    }
+}
